Remove unneeded shellVariable definitions
Defaults for the DEFAULT_SESSION and START_SERVER_COMMAND vars are
handled in the bundle code, there's no need to define them in bundle
preferences -- especially when the start command was broken out-of-the-
box because ipy_textmateserver is not distributed with current IPython!!

The user can set overrides in TM preferences as is common convention
instead of being asked to open the Bundle Editor. Doc update will be
forthcoming.